style: show changelog entry date on the side
This commit makes use of docusaurus swizzling to override the blog item component.
The layout now displays two columns, a sticky one showing the date of publication and another one with the content of the entry.

This might need some rework when we introduce a real blog section that will use a blog layout as well, as we'll need to create a different layout.