Complete how to guides with token
Using command `bump deploy` requires to use a parameter `token`.

This PR adapts sections How to start with ApiPlatform, and FastApi.

On the way back from ApiPlatform, I'd love to write more content
about a better usecase to describe connexion between ApiPlatform and bump.sh,

this is a first step 😉