help: improve sidebar performances
Building unique objects from `data/sidebar` definitions mapped onto their corresponding resources.
This way, the app doesn't need to access the list of resources for every entry in the menu.

I'm not sure this is the best way to do it but we go from a 24s building time to 9s.