guides(openapiv4): replace raw diagram into image & tweaks
This commit changes the raw definition of the Mermaid diagram into an
image (we don't yet have a lib to transform mermaid diagrams live in
the docs website).

I also took the liberty to add the `method:` keyword in the request
definition of the travel API, as it helps understanding the new
structure (at least it helped me). At the same time I added a sentence
mentioning the similar work of removing a nesting level (HTTP methods
were a parent level in openAPI v3 and is now part of the API request
object 🎊)

Let me know if that's fine for you.

P.S.: Be careful with the 🇪🇸 sun 🚲.