guides: add dates to filenames
The “last update” date displayed on each guide page is taken from the
date the hub content is being built.

In order to have a date related to the file, we need to provide the
date as a prefix of the filenames. As described in the Bridgetown
documentation: https://www.bridgetownrb.com/docs/collections#custom-collections

Closes #214