Display CTA component with a slot
Purpose of this PR is to display the CTA button in guides section
'OpenAPI specification'.

I noticed that component `Shared::BumpCta` was already included
in layout documentation.erb, as a slot named `:after` in component
`Shared::Sidebar`, but it was not displayed.

What is done in this PR:
- change condition to display this component:
favor ready-to-use boolean `resource.data.display_cta`

- render the slot in the component, with `<%= slotted :after %>`

- add basic CSS to keep space between this CTA and the sidebar links 😇
(no dinguerie I swear!)

- set data.slug to 'openapi-specification', used as `utm_content`.